pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sw360utils"
version = "0.1.0"
description = "SW360 Utils"
license = "MIT"
authors = ["Helio Chissini de Castro <heliocastro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
rich = "^13.7.1"
requests = "^2.32.3"
pyyaml = "^6.0.1"
pydantic-settings = "^2.3.4"

[tool.poetry.scripts]
sw360 = 'sw360utils.__main__:main'

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
ruff = "^0.3.2"
types-requests = "^2.32.0.20240712"
types-pyyaml = "^6.0.12.20240311"

[tool.pytest.ini_options]
    addopts = ["--import-mode=importlib"]
    log_cli = true
    log_cli_level = "INFO"
    pythonpath = "src"

[tool.pylint.messages_control]
    disable = [
        "abstract-method",
        "arguments-differ",
        "attribute-defined-outside-init",
        "blacklisted-name",
        "chained-comparison",
        "duplicate-code",
        "eval-used",
        "exec-used",
        "expression-not-assigned",
        "fixme",
        "global-statement",
        "invalid-name",
        "import-error",
        "logging-fstring-interpolation",
        "missing-docstring",
        "no-member",
        "no-name-in-module",
        "protected-access",
        "redefined-argument-from-local",
        "redefined-outer-name",
        "reimported",
        "stop-iteration-return",
        "too-few-public-methods",
        "too-many-ancestors",
        "too-many-arguments",
        "too-many-branches",
        "too-many-instance-attributes",
        "too-many-lines",
        "too-many-locals",
        "too-many-return-statements",
        "too-many-statements",
        "unexpected-special-method-signature",
        "unspecified-encoding",
    ]

[tool.mypy]
    plugins = [
        "pydantic.mypy",
    ]

    check_untyped_defs = true
    disallow_any_generics = true
    disallow_untyped_defs = true
    disallow_untyped_calls = false
    follow_imports = "normal"
    ignore_missing_imports = true
    no_implicit_reexport = true
    strict_optional = true
    warn_redundant_casts = true
    warn_unused_ignores = true

[tool.pydantic-mypy]
    init_forbid_extra = true
    init_typed = true
    warn_required_dynamic_aliases = true

[tool.ruff.mccabe]
    # Unlike Flake8, default to a complexity level of 10.
    max-complexity = 10

[tool.ruff]
    fix = true
    line-length = 120
    ignore = [
        'N802',   # function name should be lowercase
        'SIM105', # Suggest contextlib instead of try/except with pass
    ]
    extend-select = [
        "E",   # pycodestyle error
        "W",   # pycodestyle warning
        "F",   # pyflakes
        "A",   # flakes8-builtins
        "COM", # flakes8-commas
        "C4",  # flake8-comprehensions
        "Q",   # flake8-quotes
        "SIM", # flake8-simplify
        "PTH", # flake8-use-pathlib
        "I",   # isort
        "N",   # pep8 naming
        "UP",  # pyupgrade
        "S",   # bandit
    ]

    # Exclude a variety of commonly ignored directories.
    extend-exclude = [
        ".bzr",
        ".direnv",
        ".eggs",
        ".git",
        ".hg",
        ".mypy_cache",
        ".nox",
        ".pants.d",
        ".ruff_cache",
        ".svn",
        ".tox",
        ".venv",
        "__pypackages__",
        "_build",
        "buck-out",
        "build",
        "dist",
        "node_modules",
        "venv",
    ]
    per-file-ignores = {}

    # Allow unused variables when underscore-prefixed.
    dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

    # Assume Python 3.10.
    target-version = "py310"

[tool.ruff.flake8-tidy-imports]
    ban-relative-imports = "all"

[tool.ruff.lint.isort]
    required-imports = ["from __future__ import annotations"]
    # use_parentheses = true
    # line_length = 120

[tool.pyright]
    reportUnboundVariable = false

[tool.ruff.lint.pydocstyle]
convention = "google"
